setup modules, requirejs has error
