+++ public/js/common.js
@@ -1,0 +1,21 @@
+//commonjs configs for requirejs 
+
+//setup requirejs for frontend use
+//  var requirejs = require("requirejs");
+ requirejs.config({
+//    nodeRequire: require,
+    baseUrl: 'public/js',
+ });
+
+
+
+
+
+/*module.exports.model = function(){
+const messageSchema = new Schema({
+  message: String
+})
+const Message = mongoose.model('Message', messageSchema);
+module.exports = Message;
+
+}*/
