+++ public/js/formModule.js
@@ -1,16 +1,5 @@
 //define a module for requirejs usage
-//will probably have to move all mongodb stuff into its own module. use that module in index.js and in here
-
-define(['mongoose'], function(){ //this could require the module of the db being made
-//Create message schema
- const messageSchema = new Schema({
-  message: String
- })
-
- return {
-   messageObj = mongoose.model('Message', messageSchema) //in submit.js now can I fill/save this?
- }
-
-});
-
-//create a msg module that creates model. require that module inside the defined formmodule 
+define(function(require) {
+  return model = require('./schemaModule');
+   
+});
