+++ public/message.html
@@ -15,6 +15,7 @@
 <!--    <script src="/socket.io/socket.io.js"></script> -->
     <script src="./js/submit.js"></script>
     <script src= "./js/app.js"></script>
+<!-- not needed bc of common.js  <script data-main = "js/submit.js" src = "" ></script> -->
 </head>
 
   <body>
